Testing Border Styles and Added to Settings
I tested the border styles for the Home page to see what looks best with a white background.

I added a Themes page where users can select Light or Dark mode for the website. This will be linked to the Settings Panel page already on the website.

I added an article and image for the article in the slideshow